+++ samples/Ev.ServiceBus.Samples.Receiver/Startup.cs
@@ -1,9 +1,7 @@
-using System.Threading.Tasks;
 using Ev.ServiceBus.Sample.Contracts;
 using Ev.ServiceBus.Samples.Receiver.ServiceBus;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc.Filters;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
@@ -27,7 +25,9 @@ namespace Ev.ServiceBus.Samples.Receiver
                 settings =>
                 {
                     settings.WithConnection(""); // Provide a connection string here!
-                });
+                })
+                // Enables you to execute code whenever execution of a message starts, succeeded or failed
+                .RegisterEventListener<ServiceBusEventListener>();
 
             // For this sample to work, you need have an Azure service bus namespace created with the following resources:
             // - A queue named "myqueue"