+++ src/Ev.ServiceBus/Reception/DuplicateEvenTypeIdDeclarationException.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Linq;
+using Ev.ServiceBus.Abstractions;
 
 namespace Ev.ServiceBus.Reception
 {
@@ -7,13 +8,11 @@ namespace Ev.ServiceBus.Reception
     {
         public DuplicateEvenTypeIdDeclarationException(MessageReceptionRegistration[] duplicates)
         {
-            Duplicates = duplicates;
             Message = "You cannot register the same PayloadTypeId twice for the same subscription.\n"
                       + "Duplicates at fault :\n"
                       + $"{string.Join("\n", duplicates.Select(o => $"{o.Options.ClientType} {o.Options.ResourceId} => {o.PayloadTypeId} => {o.HandlerType}"))}";
         }
 
-        public MessageReceptionRegistration[] Duplicates { get; }
         public override string Message { get; }
     }
 }