+++ docs/CHANGELOG.md
@@ -10,6 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 ### Removed
 
+## 3.2.0
+- Added log messages that tells if Ev.ServiceBus is deactivated or if only reception is deactivated.
+- Ev.ServiceBus.HealthChecks is now compatible with netcore3.1
+- Refactored internal registration of dispatches and receptions.
+- Added `IServiceBusEventListener` interface that lets you hook up to the internal events `start`, `succeeded` and `failed`.
+
 ## 3.1.0
 
 - Add custom tags for health check